+++ packages/flutter_rocket/example/test/fake_rocket_client.dart
@@ -6,11 +6,11 @@ class RocketClientTest extends RocketClient {
     this.dummyData,
   ) : super(url: '');
   @override
-  Future request<T>(String endpoint,
+  Future<RocketModel> request<T>(String endpoint,
       {RocketModel<T>? model,
       HttpMethods method = HttpMethods.get,
       Function(dynamic data)? inspect,
-      List<String>? targetData,
+      List<String>? target,
       Map<String, dynamic>? data,
       Map<String, dynamic>? params}) async {
     model!.state = RocketState.loading;